backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0
greenlet==3.2.5

# Document parsing
python-docx==1.1.2
pdfplumber==0.11.4

# AI / Embeddings
sentence-transformers==3.3.1
numpy==2.0.2

# Utilities
pydantic==2.10.4
pydantic-settings==2.7.1
eval_type_backport==0.3.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
